feat: new structure for simplicity
