--- .npmrc
@@ -1,1 +1,0 @@
-auto-install-peers = true
