+++ README.md
@@ -7,7 +7,7 @@ This example shows how to use React Email with [Turborepo](https://turbo.build) 
 This monorepo includes the following apps:
 
 - `apps/web`: a [Next.js](https://nextjs.org) app
-- `packages/react-email-client`: a [react.email](https://react.email) app
+- `packages/transactional`: a package with [react.email](https://react.email) email templates
 
 ## Instructions
