+++ apps/web/package.json
@@ -16,14 +16,13 @@
   "devDependencies": {
     "@next/eslint-plugin-next": "^14.0.2",
     "@react-email/render": "^0.0.9",
-    "@repo/eslint-config": "*",
-    "@repo/react-email-client": "*",
-    "@repo/typescript-config": "*",
     "@types/eslint": "^8.44.7",
     "@types/node": "^17.0.12",
     "@types/react": "^18.0.22",
     "@types/react-dom": "^18.0.7",
+    "@vercel/style-guide": "^5.1.0",
     "eslint": "^8.53.0",
+    "transactional": "*",
     "typescript": "^5.2.2"
   }
 }
