+++ package.json
@@ -2,25 +2,30 @@
   "name": "react-email-turborepo-yarn-example",
   "private": true,
   "scripts": {
-    "build": "turbo build",
-    "dev": "turbo dev",
-    "lint": "turbo lint",
+    "build": "turbo build --filter=!react-email-client",
+    "dev": "turbo dev --filter=!react-email-client",
+    "lint": "turbo lint --filter=!react-email-client",
     "format": "prettier --write \"**/*.{ts,tsx,md}\""
   },
   "devDependencies": {
-    "@repo/eslint-config": "*",
-    "@repo/typescript-config": "*",
     "prettier": "^3.1.0",
     "turbo": "latest"
   },
   "packageManager": "yarn@3.4.1",
   "workspaces": [
     "apps/*",
+    "packages/transactional/.react-email",
     "packages/*"
   ],
   "dependencies": {
+    "@typescript-eslint/eslint-plugin": "^6.11.0",
+    "@typescript-eslint/parser": "^6.11.0",
+    "@vercel/style-guide": "^5.1.0",
+    "eslint-config-prettier": "^9.0.0",
+    "eslint-config-turbo": "^1.10.12",
     "eslint-plugin-only-warn": "^1.1.0",
     "next": "^14.0.3",
-    "styled-jsx": "^5.1.2"
+    "styled-jsx": "^5.1.2",
+    "typescript": "^5.2.2"
   }
 }
