--- packages/eslint-config/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "@repo/eslint-config",
-  "version": "0.0.0",
-  "private": true,
-  "files": [
-    "library.js",
-    "next.js",
-    "react-internal.js"
-  ],
-  "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^6.11.0",
-    "@typescript-eslint/parser": "^6.11.0",
-    "@vercel/style-guide": "^5.1.0",
-    "eslint-config-prettier": "^9.0.0",
-    "eslint-config-turbo": "^1.10.12",
-    "eslint-plugin-only-warn": "^1.1.0",
-    "typescript": "^5.2.2"
-  }
-}
