+++ packages/transactional/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "transactional",
+  "version": "0.0.0",
+  "private": true,
+  "scripts": {
+    "dev": "email dev -p 3001",
+    "export": "email export"
+  },
+  "dependencies": {
+    "@react-email/components": "0.0.12",
+    "@react-email/render": "0.0.10",
+    "react": "^18.2.0",
+    "react-email": "1.10.0"
+  },
+  "devDependencies": {
+    "@types/react": "^18.2.38"
+  }
+}
