--- packages/typescript-config/package.json
@@ -1,9 +1,0 @@
-{
-  "name": "@repo/typescript-config",
-  "version": "0.0.0",
-  "private": true,
-  "license": "MIT",
-  "publishConfig": {
-    "access": "public"
-  }
-}
